dashboards/legacy_demo/requirements.txt
streamlit
plotly
pandas
google-cloud-bigquery
db-dtypes
